Renaming Vungle ad delegate in the Vungle adapter.
PiperOrigin-RevId: 282810563
