+++ adapters/Vungle/VungleAdapter/GADMAdapterVungleRouter.h
@@ -14,51 +14,24 @@
 
 #import <Foundation/Foundation.h>
 #import <VungleSDK/VungleSDK.h>
+#import "GADMAdapterVungleDelegate.h"
 #import "VungleAdNetworkExtras.h"
-
-/// Vungle adapter ad type.
-typedef NS_ENUM(NSUInteger, GADMAdapterVungleAdType) {
-  GADMAdapterVungleAdTypeUnknown,       ///< Unknown adapter type.
-  GADMAdapterVungleAdTypeRewarded,      ///< Rewarded adapter type.
-  GADMAdapterVungleAdTypeInterstitial,  ///< Interstitial adapter type.
-  GADMAdapterVungleAdTypeBanner         ///< Banner adapter type.
-};
-
-typedef NS_ENUM(NSUInteger, BannerRouterDelegateState) {
-  BannerRouterDelegateStateRequesting,
-  BannerRouterDelegateStateCached,
-  BannerRouterDelegateStatePlaying,
-  BannerRouterDelegateStateClosing,
-  BannerRouterDelegateStateClosed
-};
-
-@protocol VungleDelegate <NSObject>
-- (void)initialized:(BOOL)isSuccess error:(nullable NSError *)error;
-- (void)adAvailable;
-- (void)adNotAvailable:(nonnull NSError *)error;
-- (void)willShowAd;
-- (void)willCloseAd:(BOOL)completedView didDownload:(BOOL)didDownload;
-- (void)didCloseAd:(BOOL)completedView didDownload:(BOOL)didDownload;
-@property(nonatomic, strong, nullable) NSString *desiredPlacement;
-@property(nonatomic, assign) GADMAdapterVungleAdType adapterAdType;
-@optional
-@property(nonatomic, assign) BannerRouterDelegateState bannerState;
-@end
 
 @interface GADMAdapterVungleRouter : NSObject <VungleSDKDelegate>
 
 + (nonnull GADMAdapterVungleRouter *)sharedInstance;
-- (void)initWithAppId:(nonnull NSString *)appId delegate:(nullable id<VungleDelegate>)delegate;
+- (void)initWithAppId:(nonnull NSString *)appId
+             delegate:(nullable id<GADMAdapterVungleDelegate>)delegate;
 - (BOOL)playAd:(nonnull UIViewController *)viewController
-      delegate:(nonnull id<VungleDelegate>)delegate
+      delegate:(nonnull id<GADMAdapterVungleDelegate>)delegate
         extras:(nullable VungleAdNetworkExtras *)extras;
 - (nullable NSError *)loadAd:(nonnull NSString *)placement
-                withDelegate:(nonnull id<VungleDelegate>)delegate;
-- (void)removeDelegate:(nonnull id<VungleDelegate>)delegate;
+                withDelegate:(nonnull id<GADMAdapterVungleDelegate>)delegate;
+- (void)removeDelegate:(nonnull id<GADMAdapterVungleDelegate>)delegate;
 - (BOOL)hasDelegateForPlacementID:(nonnull NSString *)placementID
                       adapterType:(GADMAdapterVungleAdType)adapterType;
 - (nullable UIView *)renderBannerAdInView:(nonnull UIView *)bannerView
-                                 delegate:(nonnull id<VungleDelegate>)delegate
+                                 delegate:(nonnull id<GADMAdapterVungleDelegate>)delegate
                                    extras:(nullable VungleAdNetworkExtras *)extras
                            forPlacementID:(nonnull NSString *)placementID;
 - (void)completeBannerAdViewForPlacementID:(nullable NSString *)placementID;
